商品管理　削除機能完成
