+++ src/main/java/jp/co/aforce/ManageServlet/ManageDeleteServlet.java
@@ -2,7 +2,6 @@ package jp.co.aforce.ManageServlet;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.List;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -10,8 +9,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import jp.co.aforce.beans.ManageList;
-import jp.co.aforce.dao.ManageListDAO.ManageDeleteDAO;
+import jp.co.aforce.dao.ManageDeleteDAO;
 
 @WebServlet("/manageDeleteServlet")
 public class ManageDeleteServlet extends HttpServlet {
@@ -22,8 +20,10 @@ public class ManageDeleteServlet extends HttpServlet {
 		try {
 			HttpSession session = request.getSession();
 
+			int product_id = Integer.parseInt(request.getParameter("product_id"));
+			
 			ManageDeleteDAO dao = new ManageDeleteDAO();
-			List<ManageList> manageDelete = dao.search();
+			int manageDelete = dao.delete(product_id);
 
 			session.setAttribute("manageDelete", manageDelete);
